// src/app/modules/home/components/widget/lib/gridvo-components/fd-frequently-links/enum/enum.ts
export enum DialogEmiterType {
  ADD = "_add",
  ADD_DESTORY = "_add_destory",
  MODIFY = "_modify",
  MODIFY_DESTORY = "_modify_destory",
}

export enum DialogEmiterName {
  IS_CHECK = "isCheck",
  ID = "id",
  SYSTEM_NAME = "systemName",
  LINK = "link",
  LINK_TYPE = "linkType",
  REMARKS = "remarks",
  CREATER = "creater",
  CREATE_TIME = "createTimes",
  DISABLED = "disabled",
  SLOT_NAME = "slotName",
}